part 4

+++ osa4/blogilista/controllers/blogs.js
@@ -1,36 +1,76 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
+const User = require('../models/user');
+const jwt = require('jsonwebtoken');
+
+const getToken = req => {
+    const auth = req.get('authorization');
+    if (auth && auth.toLowerCase().startsWith('bearer ')) {
+        return auth.substring(7);
+    }
+    return;
+}
 
 blogsRouter.get('/', (request, response) => {
     Blog
-        .find({})
+        .find({}).populate('user', { name: 1 })
         .then(blogs => {
             response.json(blogs.map(b => b.toJSON()))
         })
 })
 
-blogsRouter.post('/', (request, response) => {
-    const blog = new Blog(request.body)
+blogsRouter.post('/', async (request, response, next) => {
+    const token = getToken(request);
 
-    blog
-        .save()
-        .then(result => {
-            response.status(201).json(result)
-        })
-        .catch(() => {
-            response.status(400).end();
-        })
+    try {
+        const decToken = jwt.verify(token, process.env.SECRET);
+
+        if (!token || !decToken.id) {
+            return response.status(401).end();
+        }
+
+        const user = await User.findById(decToken.id);
+
+        const blog = new Blog({
+            ...request.body,
+            user: user._id
+        });
+
+        const savedBlog = await blog.save();
+        user.blogs = user.blogs.concat(blog);
+        await user.save();
+        response.json(savedBlog);
+    } catch (e) {
+        next(e)
+    }
 });
 
-blogsRouter.delete('/:id', (request, response) => {
+blogsRouter.delete('/:id', async (request, response) => {
     var id = request.params.id;
-    Blog.findByIdAndDelete(id)
-        .then(() => {
-            response.status(200).end();
-        })
-        .catch(e => {
-            response.status(500).json(e);
-        });
+    const token = getToken(request);
+
+    try {
+        const decToken = jwt.verify(token, process.env.SECRET);
+
+        if (!token || !decToken.id) {
+            return response.status(401).end();
+        }
+
+        const user = await User.findById(decToken.id);
+        const blog = await Blog.findById(id);
+
+        if (user._id.toString() === blog.user.toString()) {
+            Blog.findByIdAndDelete(id)
+                .then(() => {
+                    response.status(200).end();
+                })
+                .catch(e => {
+                    response.status(500).json(e);
+                });
+        }
+    } catch (e) {
+        next(e)
+    }
 });
 
 blogsRouter.post('/:id', (req, res) => {
@@ -38,7 +78,7 @@ blogsRouter.post('/:id', (req, res) => {
     const body = req.body;
 
     if (!body.likes) {
-        return res.status(400).json({ error: 'content missing'});
+        return res.status(400).json({ error: 'content missing' });
     }
 
     console.log(body);
@@ -51,9 +91,9 @@ blogsRouter.post('/:id', (req, res) => {
     b.save().then(s => {
         res.status(200).json(s.toJSON());
     })
-    .catch(e => {
-        res.status(500).json(e);
-    })
+        .catch(e => {
+            res.status(500).json(e);
+        })
 
 });
